requirements.txt
asgiref==3.9.1
biopython==1.85
contourpy==1.3.3
cycler==0.12.1
Django==5.2.4
fonttools==4.59.0
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
